change eslint rules
"import/prefer-default-export" is just a flavour. default exports are less flexible, if adding new exports to a module you have to change to named export anyway, so I prefer to always use named exports. It should definitely not be a hard 'error'

"import/extensions": "off" for native ES6-Modules in a browser environment this rule is strictly wrong. You can only leave out module file extensions when running in nodejs or if you use a bundler that resolves the modules filename.