Poll status for downloads (#86)
* implement repeated polling of a URL until a custom validation function returns true

* change eslint rules

"import/prefer-default-export" is just a flavour. default exports are less flexible, if adding new exports to a module you have to change to named export anyway, so I prefer to always use named exports. It should definitely not be a hard 'error'

"import/extensions": "off" for native ES6-Modules in a browser environment this rule is strictly wrong. You can only leave out module file extensions when running in nodejs or if you use a bundler that resolves the modules filename.

* ignore eslint rule no-await-in-loop

in many cases this rule does make sense to improve performance and run promises in parallel. Here it doesn't as the order of execution in our loop matters.

* apply black and isort

* change validation logic and signature of poll function and add documentation

* factored out some DOM Helpers

* fixed poll function

Before it only catched errors after the first api call, now it is checking every time

* simulate different timings for the results

* refactoring poll and create-results code to browser compatible native ES6 Modules

* Update sketch_map_tool/routes.py

Co-authored-by: Matthias Schaub <83658582+matthiasschaub@users.noreply.github.com>

* Update sketch_map_tool/routes.py

Co-authored-by: Matthias Schaub <83658582+matthiasschaub@users.noreply.github.com>

Co-authored-by: Matthias Schaub <83658582+matthiasschaub@users.noreply.github.com>