feat(frontend): add an observable property "activeLayer" to the LayerSwitcher Control
now it's possible to use an EventListener to get notified when the baselayer is switched with:
layerSwitcher.on("change:activeLayer", handleLayerSwitch)

This will be used to update the hidden form field with the layer information.
May also be used to update the location.href in the browser bar to create a bookmarkable URL with a layer state.