refactor(tasks/routes): process files in parallel
instead of georeferencing uploaded files in one celery tasks sequential
(w/ the output of a single big zip file which if too big leads to memory
issues) files are now processed in parallel in one celery group.

The same is done for digitize sketches: Digitize files in parallel as
celery group instead of as single celery task.

Other changes:
- refactor: move merge and zip_ to helpers module
- refactor(celery): use chord and immutable signature. Use chord for executing task after group is finished.