feat(map_generation): use aruco markers on sketch map
Optionally use aruco markers instead of globes on Sketch Maps.
Every Aruco marker is unique (in contrast to the globes).

Aruco markers can be used in clipping the map frame from uploaded sketch
maps using as it is done currently using BRSIK or an marking detection
algorithm provided by opencv2 in the future.